feat(comment): restructure comment module to version 1 and update imports
