+++ apps/official-journal-admin-api/src/app/case/case.module.ts
@@ -7,7 +7,7 @@ import {
   AdvertTypeModule,
   ApplicationUserController,
   ApplicationUserModule,
-  CommentModule,
+  CommentModuleV1,
   InstitutionAdminController,
   InstitutionController,
   InstitutionModule,
@@ -26,7 +26,7 @@ import { CaseController } from './case.controller'
     LoggingModule,
     SharedCaseModule,
     SharedJournalModule,
-    CommentModule,
+    CommentModuleV1,
     SignatureModule,
     AdminUserModule,
     ApplicationUserModule,
