+++ libs/shared/modules/src/application/application.service.ts
@@ -49,7 +49,7 @@ import {
 import { IAttachmentService } from '../attachments/attachment.service.interface'
 import { IAuthService } from '../auth/auth.service.interface'
 import { ICaseService } from '../case/case.module'
-import { ICommentService } from '../comment/comment.service.interface'
+import { ICommentService } from '../comment/v1/comment.service.interface'
 import { IS3Service } from '../s3/s3.service.interface'
 import { ISignatureService } from '../signature/signature.service.interface'
 import { IUtilityService } from '../utility/utility.service.interface'
