+++ libs/shared/modules/src/case/services/update/case-update.service.ts
@@ -30,8 +30,8 @@ import { BadRequestException, Inject, Injectable } from '@nestjs/common'
 import { InjectModel } from '@nestjs/sequelize'
 
 import { IApplicationService } from '../../../application/application.service.interface'
-import { ICommentService } from '../../../comment/comment.service.interface'
-import { IUtilityService } from '../../../utility/utility.service.interface'
+import { ICommentService } from '../../../comment/v1'
+import { IUtilityService } from '../../../utility/utility.module'
 import { updateCaseBodyMapper } from '../../mappers/case-update-body.mapper'
 import { CaseCategoriesModel, CaseModel } from '../../models'
 import { ICaseUpdateService } from './case-update.service.interface'
