+++ libs/shared/modules/src/comment/v1/index.ts
@@ -1,0 +1,5 @@
+export * from './migrations'
+export * from './models'
+export * from './comment.module'
+export * from './comment.service'
+export * from './comment.service.interface'
