+++ libs/shared/modules/src/index.ts
@@ -4,7 +4,7 @@ export * from './case/case.module'
 export * from './health/health.module'
 export * from './journal'
 export * from './journal/models'
-export * from './comment/comment.module'
+export * from './comment/v1/comment.module'
 export * from './utility/utility.module'
 export * from './pdf/pdf.module'
 export * from './signature/signature.module'
